chore: move date related utils to date_time_util.dart
