--- mobile/lib/utils/seconds_to_duration.dart
@@ -1,12 +1,0 @@
-/// Returns the duration in the format "h:mm:ss" or "m:ss".
-String secondsToDuration(int totalSeconds) {
-  final hours = totalSeconds ~/ 3600;
-  final minutes = (totalSeconds % 3600) ~/ 60;
-  final seconds = totalSeconds % 60;
-
-  if (hours > 0) {
-    return '${hours.toString().padLeft(1, '0')}:${minutes.toString().padLeft(2, '0')}:${seconds.toString().padLeft(2, '0')}';
-  } else {
-    return '${minutes.toString().padLeft(1, '0')}:${seconds.toString().padLeft(2, '0')}';
-  }
-}
